SAI example with wolfson wm8960 codec
Provides a sample application for playing back audio without DMA
directly writing to the FIFO of the SAI peripheral, which should result
in an I2S signaling to the WM8960 codec on the EVK boards. The codec
should then output a nice square wave on one channel and a sine wave on
the other.

The example in this case uses a PIT timer to print the status and a SAI1
interrupt noting a FIFO request to synthesize samples when needed.